pyproject.toml
[project]
name = "func-ai-foundry-agent-search"
version = "0.1.0"
description = ""
authors = [
    {name = "kimtth"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "azure-functions (>=1.23.0,<2.0.0)",
    "azure-ai-projects (>=1.0.0b12,<2.0.0)",
    "azure-identity (>=1.23.1,<2.0.0)",
    "azure-ai-agents (>=1.0.2,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "azure-cosmos (>=4.9.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"